Use a more compact NFA representation
First, given the way that the NFA is constructed from regular
expressions, all character transitions from a given state (if any) have
the same destination state. As a result, we don't need to store a full
destination set for each set of characters (only one destination set).

Second, use a vector of integers for states instead of a vector<bool> (a
bit vector) for a set of states.

This significantly speeds up compilation time.