Remove uses of range algorithms and workaround LLVM bug 63761
At this point, it seems safer to not depend on the implementation of
ranges that depend on _LIBCPP_ENABLE_EXPERIMENTAL in libc++ (aside from
type definitions that are always provided). Use the old-style iterator
versions instead.

Also, add a workaround for LLVM (Clang) bug 63761, which will be
triggered by calls to vector::insert. The bug is a frontend crash
specific to libc++. This makes the algorithm less efficient,
unfortunately.

Adjust the test case to test both the ranges and the iterator interface
for the scanner.